requirements.txt
matplotlib==3.10.0
pandas==2.2.3
Pillow==11.1.0
torch==2.6.0
torchvision==0.21.0
transformers==4.48.3
